Move detect files
