+++ packages/extensions/core/lib/configuration/index.ts
@@ -3,3 +3,4 @@ export * from "./auto-rest-configuration-impl";
 export * from "./configuration-view";
 export * from "./message-emitter";
 export * from "./directive";
+export * from "./configuration-file-resolver";
